admin files replaced & tests replaced
