+++ employee/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import IssueLog, Employee
+
+admin.site.register(Employee)
+admin.site.register(IssueLog)
