+++ project/admin.py
@@ -1,3 +1,9 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Project, Sprint, Issue, ProjectTeam
+
+admin.site.register(Project)
+admin.site.register(Sprint)
+admin.site.register(Issue)
+admin.site.register(ProjectTeam)
+
